拆分Circle
