+++ src/manimgeo/components/__init__.py
@@ -2,4 +2,4 @@ from ..components.base import GeometryAdapter, BaseGeometry
 from .point import Point
 from .line import Line, LineSegment, Ray, InfinityLine
 from .vector.vector import *
-from .circle.circle import *
+from .circle import Circle
