+++ src/manimgeo/components/circle/construct.py
@@ -1,0 +1,7 @@
+from typing import Union, Literal
+
+CircleConstructType = Literal[
+    "PR", "PP", "L", "PPP", "TranslationCirV",
+    "InverseCirCir", "InscribePPP"
+]
+Number = Union[float, int]
